Refactor Room Generator (DesignCanvas) structure and logic
- Extracted property logic to domain service
- Extracted canvas interaction to custom hook
- Split DesignCanvasRender into sub-components
- Added documentation for the refactoring
